# Proof of concept for faster start time.