resolves https://dbtlabs.atlassian.net/browse/MNTL-180

### Description

Proof of concept for faster startup time. Delays slow custom compilation steps until they are needed, if they are needed.

**Before:**

dbt 1.913s
dbt --version 2.184s

**After:**

dbt 0.674s
dbt --version 1.150s


### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-common/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] I have run this code in development and it appears to resolve the stated issue
- [x] I have [opened an issue to add/update docs](https://github.com/dbt-labs/docs.getdbt.com/issues/new/choose), or docs changes are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-common/blob/main/CONTRIBUTING.md#Adding-CHANGELOG-Entry)